schools 폴더에서 feeds 를 분리
